chore: re-arranged the code in the sessions module to prepare for the client/daemon split
