+++ crates/sessions/src/core/mod.rs
@@ -1,0 +1,7 @@
+pub mod decision;
+pub mod expansion;
+pub mod lifecyclehook;
+pub mod loadout;
+pub mod policy;
+pub mod primitives;
+pub mod source;
