+++ crates/sessions/src/wire/mod.rs
@@ -1,0 +1,1 @@
+// wire types added in step 2
